pyproject.toml
[project]
name = "whisper"
version = "0.1.0"
description = "Whisper ASR Service using OpenAI Whisper source"
requires-python = ">=3.12"
dependencies = [
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "fastapi",
    "uvicorn",
    "python-multipart",
]
